Create main.c
To be expanded to include an automatic WiFi config, and if not possible, use a WiFi manager to connect to an available WiFi access point. Config Serial to send debugging information from pins 1 and 3 making them available for debugger output. If debugger output not required, pins 1 and 3 can be used for other GPIO activity, such as Dallas 1 wire or DHT-11/DHT-22/BME-280 or other sensors. Good luck Squirrel